+++ packages/dubbo/src/protocol-grpc/parse-timeout.spec.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { parseTimeout } from "./parse-timeout.js";
 
 describe("parseTimeout()", function () {
