+++ packages/dubbo/src/protocol-grpc/trailer-status.spec.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import { findTrailerError, setTrailerStatus } from "./trailer-status.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { Code } from "../code.js";
 import { Int32Value } from "@bufbuild/protobuf";
