+++ packages/dubbo/src/protocol-grpc/trailer-status.ts
@@ -14,7 +14,7 @@
 
 import { Any, Message } from "@bufbuild/protobuf";
 import { Status } from "./gen/status_pb.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { decodeBinaryHeader, encodeBinaryHeader } from "../http-headers.js";
 import { Code } from "../code.js";
 import {
