+++ packages/dubbo/src/protocol-grpc/transport.ts
@@ -23,7 +23,7 @@ import { requestHeaderWithCompression } from "./request-header.js";
 import { validateResponseWithCompression } from "./validate-response.js";
 import { validateTrailer } from "./validate-trailer.js";
 import { Code } from "../code.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import type {
   UnaryResponse,
   UnaryRequest,
