+++ packages/dubbo/src/protocol-triple/error-json.spec.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { Code } from "../code.js";
 import { Message, proto3, protoBase64, ScalarType } from "@bufbuild/protobuf";
 import { errorFromJson, errorToJson } from "./error-json.js";
