+++ packages/dubbo/src/protocol-triple/error-json.ts
@@ -19,7 +19,7 @@ import type {
   JsonWriteOptions,
 } from "@bufbuild/protobuf";
 import { Code } from "../code.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { codeFromString, codeToString } from "./code-string.js";
 
 /**
