+++ packages/dubbo/src/protocol-triple/transport.ts
@@ -29,7 +29,7 @@ import { createEndStreamSerialization, endStreamFlag } from "./end-stream.js";
 import { transformConnectPostToGetRequest } from "./get-request.js";
 import type { CommonTransportOptions } from "../protocol/transport-options.js";
 import { Code } from "../code.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { appendHeaders } from "../http-headers.js";
 import type {
   UnaryResponse,
