+++ packages/dubbo/src/protocol-triple/version.ts
@@ -14,7 +14,7 @@
 
 import { headerProtocolVersion } from "./headers.js";
 import { paramConnectVersion } from "./query-params.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { Code } from "../code.js";
 
 /**
