+++ packages/dubbo/src/protocol/compression.spec.ts
@@ -14,7 +14,7 @@
 
 import { compressionNegotiate } from "./compression.js";
 import type { Compression } from "./compression.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { node16FetchHeadersPolyfill } from "../node16-polyfill-helper.spec.js";
 
 node16FetchHeadersPolyfill();
