+++ packages/dubbo/src/protocol/envelope.spec.ts
@@ -20,7 +20,7 @@ import {
   envelopeDecompress,
 } from "./envelope.js";
 import type { Compression } from "./compression.js";
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { Code } from "../code.js";
 import { node16WhatwgStreamPolyfill } from "../node16-polyfill-helper.spec.js";
