+++ packages/dubbo/src/protocol/envelope.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ConnectError } from "../connect-error.js";
+import { ConnectError } from "../dubbo-error.js";
 import { Code } from "../code.js";
 import { compressedFlag } from "./compression.js";
 import type { Compression } from "./compression.js";
