+++ packages/dubbo/src/router.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import type { MethodInfo, ServiceType } from "@bufbuild/protobuf";
-import { ConnectError } from "./connect-error.js";
+import { ConnectError } from "./dubbo-error.js";
 import { Code } from "./code.js";
 import {
   createMethodImplSpec,
